Add check for file before cleanup

+++ backup-mysql.sh
@@ -23,7 +23,9 @@ docker exec ${CONTAINER} /usr/bin/mysqldump --single-transaction -u ${USERNAME} 
 
 if [ $? -ne 0 ]; then
 	echo "[$0] FATAL: Failed to create backupfile ${FILENAME}"
-	rm ${FILENAME}
+	if [ -f ${FILENAME} ]; then
+		rm ${FILENAME}
+	fi
 	exit 1
 fi
